backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
requests
python-dotenv
PyMySQL
marshmallow
Flask-Cors
PyJWT